+++ lib/controller.dart
@@ -26,7 +26,7 @@ export 'package:mvc_pattern/mvc_pattern.dart' show AppMVC, AppConMVC;
 
 //Controller
 export 'package:mvc_application/src/controller/app.dart'
-    show AppController, ControllerMVC;
+    show AppController, ErrorHandler, ControllerMVC;
 
 export 'package:mvc_application/src/controller/deviceinfo.dart' show DeviceInfo;